backend/app/tests/utils/utils.py
import random
import string

import pytest
from fastapi.testclient import TestClient


def random_lower_string() -> str:
    return "".join(random.choices(string.ascii_lowercase, k=32))


def random_email() -> str:
    return f"{random_lower_string()}@{random_lower_string()}.com"


def get_superuser_token_headers(client: TestClient) -> dict[str, str]:
    # Authentication removed. Tests that relied on auth were deleted.
    pytest.skip("auth helpers removed")
